# Conflicts:
#	app/src/com/constellio/app/extensions/api/scripts/ResetParsedContentScript.java

+++ app/src/com/constellio/app/extensions/api/scripts/ResetParsedContentScript.java
@@ -3,7 +3,6 @@ package com.constellio.app.extensions.api.scripts;
 import com.constellio.app.modules.rm.services.RMSchemasRecordsServices;
 import com.constellio.app.modules.rm.wrappers.Document;
 import com.constellio.app.services.factories.AppLayerFactory;
-import com.constellio.app.services.migrations.ConstellioEIM;
 import com.constellio.data.dao.services.bigVault.SearchResponseIterator;
 import com.constellio.data.io.services.facades.IOServices;
 import com.constellio.model.entities.records.Record;
@@ -11,7 +10,6 @@ import com.constellio.model.entities.records.Transaction;
 import com.constellio.model.entities.records.wrappers.UserDocument;
 import com.constellio.model.entities.schemas.Schemas;
 import com.constellio.model.services.contents.ContentManager;
-import com.constellio.model.services.records.SchemasRecordsServices;
 import com.constellio.model.services.search.SearchServices;
 import com.constellio.model.services.search.query.logical.LogicalSearchQuery;
 import com.constellio.model.services.search.query.logical.LogicalSearchQueryOperators;
@@ -39,15 +37,15 @@ public class ResetParsedContentScript extends ScriptWithLogOutput {
 
         for(String collection : appLayerFactory
                 .getCollectionsManager().getCollectionCodesExcludingSystem()) {
-            RMSchemasRecordsServices schemasRecordsServices = new RMSchemasRecordsServices(collection, modelLayerFactory);
+            RMSchemasRecordsServices rmSchemasRecordsServices = new RMSchemasRecordsServices(collection, appLayerFactory);
             LogicalSearchQuery logicalSearchQuery = new LogicalSearchQuery();
-            logicalSearchQuery.setCondition(LogicalSearchQueryOperators.from(schemasRecordsServices.userDocument.schemaType()).returnAll());
+            logicalSearchQuery.setCondition(LogicalSearchQueryOperators.from(rmSchemasRecordsServices.documentSchemaType()).returnAll());
             SearchResponseIterator<Record> searchResponseIterator = searchServices.recordsIterator(logicalSearchQuery, 500);
             ContentManager contentManager = modelLayerFactory.getContentManager();
 
             while(searchResponseIterator.hasNext()) {
                 Record currentRecord = searchResponseIterator.next();
-                Document document = schemasRecordsServices.wrapDocument(currentRecord);
+                Document document = rmSchemasRecordsServices.wrapDocument(currentRecord);
 
                 String fileName = contentManager.getParsedContentFileName(document.getContent().getCurrentVersion().getHash());
                 File parsedFilePath = contentManager.getContentDao().getFileOf(fileName);
@@ -73,12 +71,13 @@ public class ResetParsedContentScript extends ScriptWithLogOutput {
             transactions.add(currentTransaction);
         }
 
-        for(File currentFileToSuppress : documentToSuppress) {
-            ioServices.deleteQuietly(currentFileToSuppress);
-        }
-
-        for(Transaction transaction: transactions) {
+            for(File currentFileToSuppress : documentToSuppress) {
+                if (currentFileToSuppress.exists()) {ioServices.deleteQuietly(currentFileToSuppress);
+            outputLogger.appendToFile("Deleted file : " + currentFileToSuppress.getName() + "\n");
+                }}for(Transaction transaction: transactions) {
             recordServices.execute(transaction);
         }
+
+        outputLogger.appendToFile("Execution done. Deleted : " + documentToSuppress.size() + " parsed document");
     }
 }
